requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
exceptiongroup==1.1.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.0
Flask==2.3.2
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
langdetect==1.0.9
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
newspaper3k==0.2.8
nltk==3.8.1
openai==0.27.6
outcome==1.2.0
Pillow==9.5.0
pyahocorasick==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-file==1.5.1
selenium==4.9.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tinysegmenter==0.3
tldextract==3.4.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
wsproto==1.2.0
yarl==1.9.2
